AG_ReadSurface(): Initialize pf.mode correctly (fix an uninitialized read).
Prevent warnings from "enum *" -> "Uint *", "Uint" -> "AG_Size".

Prevent warnings related to implicit float <-> double conversion.

AG_Console: Avoid a close() usage under Win32.

M_BezierPrimitives: Avoid a c99 usage in M_PointSetAdd2() call.

SG: Don't assume snprintf() exists. Use Snprintf() aka AG_Snprintf().

agartest: Avoid printf(), fprintf() and exit() usages.



git-svn-id: https://dev.csoft.net/agar/trunk@10876 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ gui/hsvpal.c
@@ -274,7 +274,7 @@ AG_HSVPal_UpdateHue(AG_HSVPal *_Nonnull pal, int x, int y)
 
 	h = Atan2((float)y, (float)x);
 	if (h < 0) {
-		h += (2.0f*AG_PI);
+		h += 2.0f*(float)AG_PI;
 	}
 	AG_SetFloat(pal, "hue", h/(2.0f*AG_PI)*360.0f);
